Rename ndarray file to `wrappers.py`
